+++ Classes/FrontEnd/EventEditor.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use OliverKlee\Oelib\Email\SystemEmailFromBuilder;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\ContentObject\ContentObjectRenderer;
 
@@ -1420,7 +1421,8 @@ class Tx_Seminars_FrontEnd_EventEditor extends \Tx_Seminars_FrontEnd_Editor
             /** @var \Tx_Oelib_Mail $eMail */
             $eMail = GeneralUtility::makeInstance(\Tx_Oelib_Mail::class);
             $eMail->addRecipient($reviewer);
-            $eMail->setSender(self::getLoggedInUser());
+            $eMail->setSender($this->getEmailSender());
+            $eMail->setReplyTo(self::getLoggedInUser());
             $eMail->setSubject($this->translate('publish_event_subject'));
             $eMail->setMessage($this->createEMailContent($event));
 
@@ -1528,7 +1530,8 @@ class Tx_Seminars_FrontEnd_EventEditor extends \Tx_Seminars_FrontEnd_Editor
         /** @var \Tx_Oelib_Mail $eMail */
         $eMail = GeneralUtility::makeInstance(\Tx_Oelib_Mail::class);
         $eMail->addRecipient($reviewer);
-        $eMail->setSender(self::getLoggedInUser());
+        $eMail->setSender($this->getEmailSender());
+        $eMail->setReplyTo(self::getLoggedInUser());
         $eMail->setSubject($this->translate('save_event_subject'));
         $eMail->setMessage($this->createAdditionalEmailContent());
 
@@ -2838,5 +2841,16 @@ class Tx_Seminars_FrontEnd_EventEditor extends \Tx_Seminars_FrontEnd_Editor
     public function setSavedFormValue($key, $value)
     {
         $this->savedFormData[$key] = $value;
+    }
+
+    protected function getEmailSender(): \Tx_Oelib_Interface_MailRole
+    {
+        $systemEmailFromBuilder = GeneralUtility::makeInstance(SystemEmailFromBuilder::class);
+        if ($systemEmailFromBuilder->canBuild()) {
+            $sender = $systemEmailFromBuilder->build();
+        } else {
+            $sender = self::getLoggedInUser();
+        }
+        return $sender;
     }
 }