+++ Tests/LegacyUnit/FrontEnd/EventEditorTest.php
@@ -1833,7 +1833,7 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
     /**
      * @test
      */
-    public function sendEMailToReviewerUsesFrontEndUserNameAsFromNameForMail()
+    public function sendEMailToReviewerUsesTypo3DefaultFromNameAsFromNameForMail()
     {
         $seminarUid = $this->testingFramework->createRecord('tx_seminars_seminars');
         $this->createAndLoginUserWithReviewer();
@@ -1849,6 +1849,104 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
                 'publication_hash' => $formData['publication_hash'],
             ]
         );
+
+        $defaultFromAddress = 'system-foo@example.com';
+        $defaultFromName = 'Mr. Default';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultFromAddress;
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = $defaultFromName;
+
+        $this->subject->sendEMailToReviewer();
+
+        self::assertContains(
+            $defaultFromName,
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendEMailToReviewerUsesTypo3DefaultFromAddressAsFromAddressForMail()
+    {
+        $seminarUid = $this->testingFramework->createRecord('tx_seminars_seminars');
+        $this->createAndLoginUserWithReviewer();
+
+        $this->subject->setObjectUid($seminarUid);
+        $formData = $this->subject->modifyDataToInsert([]);
+
+        $this->testingFramework->changeRecord(
+            'tx_seminars_seminars',
+            $seminarUid,
+            [
+                'hidden' => 1,
+                'publication_hash' => $formData['publication_hash'],
+            ]
+        );
+
+        $defaultFromAddress = 'system-foo@example.com';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultFromAddress;
+
+        $this->subject->sendEMailToReviewer();
+
+        self::assertArrayHasKey(
+            $defaultFromAddress,
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendEMailToReviewerUsesFrontEndUserAsReplyToForMail()
+    {
+        $seminarUid = $this->testingFramework->createRecord('tx_seminars_seminars');
+        $this->createAndLoginUserWithReviewer();
+
+        $this->subject->setObjectUid($seminarUid);
+        $formData = $this->subject->modifyDataToInsert([]);
+
+        $this->testingFramework->changeRecord(
+            'tx_seminars_seminars',
+            $seminarUid,
+            [
+                'hidden' => 1,
+                'publication_hash' => $formData['publication_hash'],
+            ]
+        );
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = 'system-foo@example.com';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = 'Mr. Default';
+
+        $this->subject->sendEMailToReviewer();
+
+        self::assertSame(
+            ['mail@foo.com' => 'Mr. Bar'],
+            $this->mailer->getFirstSentEmail()->getReplyTo()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendEMailToReviewerWithoutTypo3DefaultFromAdressAndNameUsesFrontEndUserNameAsFromNameForMail()
+    {
+        $seminarUid = $this->testingFramework->createRecord('tx_seminars_seminars');
+        $this->createAndLoginUserWithReviewer();
+
+        $this->subject->setObjectUid($seminarUid);
+        $formData = $this->subject->modifyDataToInsert([]);
+
+        $this->testingFramework->changeRecord(
+            'tx_seminars_seminars',
+            $seminarUid,
+            [
+                'hidden' => 1,
+                'publication_hash' => $formData['publication_hash'],
+            ]
+        );
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
 
         $this->subject->sendEMailToReviewer();
 
@@ -1861,7 +1959,7 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
     /**
      * @test
      */
-    public function sendEMailToReviewerUsesFrontEndUserMailAddressAsFromAddressForMail()
+    public function sendEMailToReviewerWithoutTypo3DefaultFromAddressUsesFrontEndUserMailAddressAsFromAddressForMail()
     {
         $seminarUid = $this->testingFramework->createRecord('tx_seminars_seminars');
         $this->createAndLoginUserWithReviewer();
@@ -1877,6 +1975,9 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
                 'publication_hash' => $formData['publication_hash'],
             ]
         );
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
 
         $this->subject->sendEMailToReviewer();
 
@@ -1957,10 +2058,83 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
     /**
      * @test
      */
-    public function sendAdditionalNotificationEmailToReviewerUsesFrontEndUserAsFromName()
+    public function sendAdditionalNotificationEmailToReviewerUsesTypo3DefaultFromNameAsFromName()
     {
         $this->configuration->setAsBoolean('sendAdditionalNotificationEmailInFrontEndEditor', true);
         $this->createAndLoginUserWithReviewer();
+
+        $defaultMailFromAddress = 'system-foo@example.com';
+        $defaultMailFromName = 'Mr. Default';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultMailFromAddress;
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = $defaultMailFromName;
+
+        $this->subject->sendAdditionalNotificationEmailToReviewer();
+
+        self::assertNotNull(
+            $this->mailer->getFirstSentEmail()
+        );
+        self::assertContains(
+            $defaultMailFromName,
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendAdditionalNotificationEmailToReviewerUsesTypo3DefaultFromAddressAsFromAddress()
+    {
+        $this->configuration->setAsBoolean('sendAdditionalNotificationEmailInFrontEndEditor', true);
+        $this->createAndLoginUserWithReviewer();
+
+        $defaultMailFromAddress = 'system-foo@example.com';
+        $defaultMailFromName = 'Mr. Default';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultMailFromAddress;
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = $defaultMailFromName;
+
+        $this->subject->sendAdditionalNotificationEmailToReviewer();
+
+        self::assertNotNull(
+            $this->mailer->getFirstSentEmail()
+        );
+        self::assertArrayHasKey(
+            $defaultMailFromAddress,
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendAdditionalNotificationEmailToReviewerUsesFrontEndUserMailAsReplyTo()
+    {
+        $this->configuration->setAsBoolean('sendAdditionalNotificationEmailInFrontEndEditor', true);
+        $this->createAndLoginUserWithReviewer();
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = 'system-foo@example.com';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = 'Mr. Default';
+
+        $this->subject->sendAdditionalNotificationEmailToReviewer();
+
+        self::assertNotNull(
+            $this->mailer->getFirstSentEmail()
+        );
+        self::assertSame(
+            ['mail@foo.com' => 'Mr. Bar'],
+            $this->mailer->getFirstSentEmail()->getReplyTo()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendAdditionalNotificationEmailToReviewerWithoutTypo3DefaultFromAddressUsesFrontEndUserAsFromName()
+    {
+        $this->configuration->setAsBoolean('sendAdditionalNotificationEmailInFrontEndEditor', true);
+        $this->createAndLoginUserWithReviewer();
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
 
         $this->subject->sendAdditionalNotificationEmailToReviewer();
 
@@ -1976,10 +2150,13 @@ class Tx_Seminars_Tests_Unit_FrontEnd_EventEditorTest extends TestCase
     /**
      * @test
      */
-    public function sendAdditionalNotificationEmailToReviewerUsesFrontEndUserAsFromAddress()
+    public function sendAdditionalNotificationEmailToReviewerWithoutTypo3DefaultFromAddressUsesFrontEndUserAsFromAddress()
     {
         $this->configuration->setAsBoolean('sendAdditionalNotificationEmailInFrontEndEditor', true);
         $this->createAndLoginUserWithReviewer();
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
 
         $this->subject->sendAdditionalNotificationEmailToReviewer();