+++ Tests/LegacyUnit/SchedulerTasks/MailNotifierTest.php
@@ -1035,7 +1035,7 @@ class MailNotifierTest extends TestCase
     /**
      * @test
      */
-    public function sendRemindersToOrganizersSendsEmailWithOrganizerAsSender()
+    public function sendRemindersToOrganizersSendsEmailWithTypo3DefaultFromAddressAsSender()
     {
         $this->createSeminarWithOrganizer(
             [
@@ -1044,10 +1044,15 @@ class MailNotifierTest extends TestCase
             ]
         );
 
+        $defaultMailFromAddress = 'system-foo@example.com';
+        $defaultMailFromName = 'Mr. Default';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultMailFromAddress;
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = $defaultMailFromName;
+
         $this->subject->sendEventTakesPlaceReminders();
 
         self::assertArrayHasKey(
-            'MrTest@example.com',
+            $defaultMailFromAddress,
             $this->mailer->getFirstSentEmail()->getFrom()
         );
     }
@@ -1055,7 +1060,7 @@ class MailNotifierTest extends TestCase
     /**
      * @test
      */
-    public function sendRemindersToOrganizersForEventWithTwoOrganizersSendsEmailWithFirstOrganizerAsSender()
+    public function sendRemindersToOrganizersForEventWithTwoOrganizersSendsEmailWithTypo3DefaultFromAddressAsSender()
     {
         $eventUid = $this->createSeminarWithOrganizer(
             [
@@ -1069,6 +1074,92 @@ class MailNotifierTest extends TestCase
         );
         $this->testingFramework->createRelation('tx_seminars_seminars_organizers_mm', $eventUid, $organizerUid);
         $this->testingFramework->changeRecord('tx_seminars_seminars', $eventUid, ['organizers' => 2]);
+
+        $defaultMailFromAddress = 'system-foo@example.com';
+        $defaultMailFromName = 'Mr. Default';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = $defaultMailFromAddress;
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = $defaultMailFromName;
+
+        $this->subject->sendEventTakesPlaceReminders();
+
+        self::assertArrayHasKey(
+            $defaultMailFromAddress,
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendRemindersToOrganizersForEventWithTwoOrganizersSendsEmailWithFirstOrganizerAsReplyTo()
+    {
+        $eventUid = $this->createSeminarWithOrganizer(
+            [
+                'begin_date' => $GLOBALS['SIM_EXEC_TIME'] + \Tx_Oelib_Time::SECONDS_PER_DAY,
+                'cancelled' => \Tx_Seminars_Model_Event::STATUS_CONFIRMED,
+            ]
+        );
+        $organizerUid = $this->testingFramework->createRecord(
+            'tx_seminars_organizers',
+            ['title' => 'foo', 'email' => 'foo@example.com']
+        );
+        $this->testingFramework->createRelation('tx_seminars_seminars_organizers_mm', $eventUid, $organizerUid);
+        $this->testingFramework->changeRecord('tx_seminars_seminars', $eventUid, ['organizers' => 2]);
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = 'system-foo@example.com';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = 'Mr. Default';
+
+        $this->subject->sendEventTakesPlaceReminders();
+
+        self::assertSame(
+            ['MrTest@example.com' => 'Mr. Test'],
+            $this->mailer->getFirstSentEmail()->getReplyTo()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendRemindersToOrganizersWithoutTypo3DefaultFromAddressSendsEmailWithOrganizerAsSender()
+    {
+        $this->createSeminarWithOrganizer(
+            [
+                'begin_date' => $GLOBALS['SIM_EXEC_TIME'] + \Tx_Oelib_Time::SECONDS_PER_DAY,
+                'cancelled' => \Tx_Seminars_Model_Event::STATUS_CONFIRMED,
+            ]
+        );
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
+
+        $this->subject->sendEventTakesPlaceReminders();
+
+        self::assertArrayHasKey(
+            'MrTest@example.com',
+            $this->mailer->getFirstSentEmail()->getFrom()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function sendRemindersToOrganizersForEventWithTwoOrganizersWithoutTypo3DefaultFromAddressSendsEmailWithFirstOrganizerAsSender()
+    {
+        $eventUid = $this->createSeminarWithOrganizer(
+            [
+                'begin_date' => $GLOBALS['SIM_EXEC_TIME'] + \Tx_Oelib_Time::SECONDS_PER_DAY,
+                'cancelled' => \Tx_Seminars_Model_Event::STATUS_CONFIRMED,
+            ]
+        );
+        $organizerUid = $this->testingFramework->createRecord(
+            'tx_seminars_organizers',
+            ['title' => 'foo', 'email' => 'foo@example.com']
+        );
+        $this->testingFramework->createRelation('tx_seminars_seminars_organizers_mm', $eventUid, $organizerUid);
+        $this->testingFramework->changeRecord('tx_seminars_seminars', $eventUid, ['organizers' => 2]);
+
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
 
         $this->subject->sendEventTakesPlaceReminders();