+++ Tests/Unit/Model/EventTest.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace OliverKlee\Seminars\Tests\Unit\Model;
 
 use Nimut\TestingFramework\TestCase\UnitTestCase;
+use OliverKlee\Oelib\Email\SystemEmailFromBuilder;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * Test case.
@@ -18,9 +20,34 @@ class EventTest extends UnitTestCase
      */
     private $subject = null;
 
+    /**
+     * @var \Tx_Seminars_Model_Organizer
+     */
+    private $organizer = null;
+
     protected function setUp()
     {
+        $this->organizer = new \Tx_Seminars_Model_Organizer();
+        $this->organizer->setData(
+            [
+                'title' => 'Brain Gourmets',
+                'email' => 'organizer@example.com',
+                'email_footer' => 'Best workshops in town!',
+            ]
+        );
+        $organizers = new \Tx_Oelib_List();
+        $organizers->add($this->organizer);
+
         $this->subject = new \Tx_Seminars_Model_Event();
+        $this->subject->setData(
+            [
+                'title' => 'A nice event',
+                'begin_date' => mktime(10, 0, 0, 4, 8, 2020),
+                'end_date' => mktime(18, 30, 0, 4, 20, 2020),
+                'registrations' => new \Tx_Oelib_List(),
+                'organizers' => $organizers,
+            ]
+        );
     }
 
     /**
@@ -29,5 +56,34 @@ class EventTest extends UnitTestCase
     public function isTitled()
     {
         self::assertInstanceOf(\Tx_Seminars_Interface_Titled::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function getEmailSenderReturnsSystemEmailMailRole()
+    {
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = 'system-foo@example.com';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = 'Mr. Default';
+        $systemEmailFromBuilder = GeneralUtility::makeInstance(SystemEmailFromBuilder::class);
+
+        self::assertEquals(
+            $systemEmailFromBuilder->build(),
+            $this->subject->getEmailSender()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function getEmailSenderReturnsFirstOrganizerMailRole()
+    {
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromAddress'] = '';
+        $GLOBALS['TYPO3_CONF_VARS']['MAIL']['defaultMailFromName'] = '';
+
+        self::assertEquals(
+            $this->organizer,
+            $this->subject->getEmailSender()
+        );
     }
 }